[flue] Extract code-review diff/state helpers into lib/code-review-diff.ts and lib/code-review-state.ts
